move distributed_test_base
