+++ apex/contrib/test/optimizers/test_dist_adam.py
@@ -5,7 +5,7 @@ import os
 import torch
 from torch.testing._internal import common_utils
 from apex.contrib.optimizers.distributed_fused_adam import DistributedFusedAdam
-from apex.transformer.testing.distributed_test_base import NcclDistributedTestBase
+from apex.testing.distributed_test_base import NcclDistributedTestBase
 
 class SimpleModel(torch.nn.Module):
     def __init__(self, num_layers, size):
